backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyscopg2-binary
pydantic
python-dotenv
requests
